Refactor social login and remove redundant components
This commit includes a major refactoring of the social login process. The `Uri` property and `OnAfterRenderAsync` method were removed from `GoogleButton.razor` and replaced with a `LoginAsync` method. The `LoginConfirmation.razor` file was simplified, and the `OnInitialized` method in `SocialAuthentication.razor` was replaced with `OnAfterRenderAsync`.

The `SocialButton.razor` and `SocialLogin.razor` files and their corresponding CSS files were removed and replaced with new versions. The `Router` component in `SocialAuthenticationRouter.razor` was updated, and a `NotFound` section was added.

New packages were added to `Rystem.Authentication.Social.Blazor.csproj`, and `ServiceCollectionExtensions.cs` saw changes to the `SocialLoginManager` and `SocialLoginLocalStorageService` classes. The `RedirectDomain` for Google was updated in `appsettings.json`, and a `User` property was added to `Home.razor`.

Finally, the `UseSocialLogin` method in `Program.cs` was replaced with calls to `MapRazorComponents` and `AddInteractiveServerRenderMode`.